Move adi transformation converters into `adi/`
